resume_demo/requirements.txt
alembic==1.12.0
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
configparser==6.0.0
cryptography==41.0.7
decorator==5.1.1
distlib==0.3.7
Django==5.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
filelock==3.12.2
firebase==4.0.1
firebase-admin==6.3.0
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.2.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.7
mysql-connector-python==8.1.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pillow==10.2.0
platformdirs==3.10.0
proto-plus==1.23.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==2.0.25
sqlparse==0.4.4
stripe==7.7.0
tweepy==4.14.0
typing_extensions==4.7.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.4
validators==0.20.0
virtualenv==20.24.2
Werkzeug==2.0.0
WTForms==3.0.1
django-ckeditor==6.7.0
